+++ ProfileFever/ProfileFever.psm1
@@ -3,9 +3,20 @@
         Root module file.
 
     .DESCRIPTION
-        The root module file loads all classes, helpers and functions into the
-        module context.
+        The root module file loads all functions and helpers into the module
+        context.
 #>
+
+[CmdletBinding()]
+param
+(
+    # Enable debug mode for the module. This will allow to debug the module
+    # functions and helpers using breakpoints but will slow down module loading
+    # due to the slow dot-sourcing.
+    [Parameter(Mandatory = $false)]
+    [System.Boolean]
+    $DebugModule = $false
+)
 
 
 ## Module Core
@@ -15,23 +26,32 @@ Set-StrictMode -Version 'Latest'
 $Script:ErrorActionPreference = 'Stop'
 $Script:ProgressPreference    = 'SilentlyContinue'
 
+
 # Module metadata
-$Script:PSModulePath = [System.IO.Path]::GetDirectoryName($PSCommandPath)
-$Script:PSModuleName = [System.IO.Path]::GetFileName($PSCommandPath).Split('.')[0]
+$Script:PSModulePath    = [System.IO.Path]::GetDirectoryName($PSCommandPath)
+$Script:PSModuleName    = [System.IO.Path]::GetFileName($PSCommandPath).Split('.')[0]
+$Script:PSModuleVersion = (Import-PowerShellDataFile -Path "$Script:PSModulePath\$Script:PSModuleName.psd1")['ModuleVersion']
 
 
 ## Module Loader
 
-# Get and dot source all helper functions (internal)
-Get-ChildItem -Path "$Script:PSModulePath\Helpers" -Filter '*.ps1' -File -Recurse |
-    ForEach-Object { . $_.FullName }
+# Get and dot source all functions
+Get-ChildItem -Path "$Script:PSModulePath\Helpers", "$Script:PSModulePath\Functions" -Filter '*.ps1' -File -Recurse |
+    ForEach-Object {
+        if ($DebugModule -or $Env:PWSH_DEBUG_MODULE -eq 'true')
+        {
+            . $_.FullName
+        }
+        else
+        {
+            . ([System.Management.Automation.ScriptBlock]::Create(
+                [System.IO.File]::ReadAllText($_.FullName, [System.Text.Encoding]::UTF8)
+            ))
+        }
+    }
 
-# Get and dot source all external functions (public)
-Get-ChildItem -Path "$Script:PSModulePath\Functions" -Filter '*.ps1' -File -Recurse |
-    ForEach-Object { . $_.FullName }
 
-
-## Module configuration
+## Module Context
 
 # Initialize all relevant launcher variables
 $Script:LAUNCHER_PATH             = "$Env:AppData\PowerShell\ProfileFever"