+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Changed: Optimize module loading time by change the dot-sourcing mode
+
 ## 5.3.1 - 2026-02-12
 
 * Fixed: Fix path issue with the VS Code workspace update